Address CRs Signed-off-by: Roger Standridge <9526806+archie2x@users.noreply.github.com>
